check performance with sql

+++ src/day7/trainningJava/src/main/java/com/beetech/trainningJava/model/PageModel.java
@@ -20,5 +20,5 @@ public class PageModel<T> {
 
     private int pageNumber;
 
-    private int totalPages;
+    private long totalPages;
 }
